+++ lib/aba_payway.dart
@@ -1,8 +1,9 @@
 library;
 
 export 'src/aba_payway.dart';
-export 'src/models/partner/get_merchant_response_data_model.dart';
-export 'src/models/partner/get_merchant_response_model.dart';
+export 'src/models/partner/inquiry_merchant_by_key_request_model.dart';
+export 'src/models/partner/inquiry_merchant_response_data_model.dart';
+export 'src/models/partner/inquiry_merchant_response_model.dart';
 export 'src/models/partner/register_merchant_request_model.dart';
 export 'src/models/partner/register_merchant_response_model.dart';
 export 'src/models/partner/response_status_model.dart';