+++ lib/src/services/partner/partner_service.dart
@@ -1,4 +1,5 @@
-import 'package:aba_payway/src/models/partner/get_merchant_response_model.dart';
+import 'package:aba_payway/src/models/partner/inquiry_merchant_by_key_request_model.dart';
+import 'package:aba_payway/src/models/partner/inquiry_merchant_response_model.dart';
 
 import '../../models/partner/register_merchant_request_model.dart';
 import '../../models/partner/register_merchant_response_model.dart';
@@ -8,7 +9,11 @@ abstract class ABAPartnerService {
     throw UnimplementedError();
   }
 
-  Future<ABAGetMerchantResponseModel> getMerchantByRegisterRef(String registerRef) async {
+  Future<ABAInquiryMerchantResponseModel> inquiryMerchantByRef(String registerRef) async {
+    throw UnimplementedError();
+  }
+
+  Future<ABAInquiryMerchantResponseModel> inquiryMerchantByKey(ABAInquiryMerchantByKeyRequestModel request) async {
     throw UnimplementedError();
   }
 }